# Transpile simple choice-type regular expressions into lists of choices to use with string replace multi